# Optimize ARM utf16 validation