Closes #136

Before:

```
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 1
input detected as UTF16 little-endian
current system detected as arm64
===========================
validate_utf16+arm64, input size: 65542, iterations: 10000, dataset: unicode_lipsum/lipsum/Emoji-Lipsum.utf16.txt
kpc_set_config failed, run the program with sudo
   1.873 GB/s (7.2 %)    0.936 Gc/s     2.00 byte/char 
validate_utf16+fallback, input size: 65542, iterations: 10000, dataset: unicode_lipsum/lipsum/Emoji-Lipsum.utf16.txt
   2.579 GB/s (3.4 %)    1.289 Gc/s     2.00 byte/char 
```

After:

```
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 1
input detected as UTF16 little-endian
current system detected as arm64
===========================
validate_utf16+arm64, input size: 65542, iterations: 10000, dataset: unicode_lipsum/lipsum/Emoji-Lipsum.utf16.txt
kpc_set_config failed, run the program with sudo
  14.304 GB/s (20.4 %)    7.152 Gc/s     2.00 byte/char 
WARNING: Measurements are noisy, try increasing iteration count (-I).
validate_utf16+fallback, input size: 65542, iterations: 10000, dataset: unicode_lipsum/lipsum/Emoji-Lipsum.utf16.txt
   2.579 GB/s (1.1 %)    1.289 Gc/s     2.00 byte/char
```