@danlark1 2022-07-12
I might have degraded other benchmarks a bit, will take a look. Maybe it comes down to the fact that `shrn` uses 1 unit whereas `uminp` 2 units.